# platform/docs/swagger.yaml
components:
  schemas:
...
    emiten.CreateRequest:
      properties:
        kode:
          description: Instrument code, unique across the exchange.
          example: BBNI
          type: string
        listed_shares:
          description: Publicly tradeable shares. Must be > 0.
          example: 18462169893
          format: int64
          type: integer
        nama:
          description: Company name.
          example: Bank Negara Indonesia Tbk
          type: string
        unlisted_shares:
          description: Restricted shares not available for trading. Defaults to 0.
          example: 0
          format: int64
          type: integer
      required:
      - kode
      - listed_shares
      - nama
      type: object
    emiten.EmitenDetail:
      properties:
        current_price:
          description: Most recent execution price. Null if never traded.
          example: 8050
          format: int64
          type: integer
        free_float_percentage:
          description: |-
            Share of total shares that is publicly tradeable. Null when there are no
            shares at all. Always sums to 100 with UnlistedPercentage.
          example: 99.2
          type: number
        highest_price:
          description: Highest price ever executed. Null if never traded.
          example: 8200
          format: int64
          type: integer
        ipo_price:
          description: |-
            Offering price this instrument was listed at. Null for instruments listed
            before the field existed.
          example: 1000
          format: int64
          type: integer
        is_active:
          description: Whether the instrument accepts new orders.
          example: true
          type: boolean
        kode:
          example: BBCA
          type: string
        listed_shares:
          example: 123275050000
          format: int64
          type: integer
        lowest_price:
          description: Lowest price ever executed. Null if never traded.
          example: 7800
          format: int64
          type: integer
        nama:
          example: Bank Central Asia Tbk
          type: string
        price_source:
          description: |-
            Where reference_price came from: "trade" or "ipo". Omitted when there is no
            reference price at all.
          enum:
          - trade
          - ipo
          example: trade
          type: string
        reference_price:
          description: |-
            Price backing the valuation: the last trade, or ipo_price until the
            instrument first trades. Null only if it has neither.
          example: 8050
          format: int64
          type: integer
        total_shares:
          example: 124275050000
          format: int64
          type: integer
        unlisted_percentage:
          description: Share of total shares that is restricted. Null when there are
            no shares.
          example: 0.8
          type: number
        unlisted_shares:
          example: 1000000000
          format: int64
          type: integer
        value:
          description: |-
            Market capitalisation: reference price × total shares. Null only when the
            instrument has neither a trade nor an IPO price.
          example: 1000414152500000
          format: int64
          type: integer
      type: object
    emiten.EmitenView:
      properties:
        is_active:
          description: |-
            Whether the instrument accepts new orders. An inactive emiten still has a
            readable book and history.
          example: true
          type: boolean
        kode:
          example: BBCA
          type: string
        listed_shares:
          example: 123275050000
          format: int64
          type: integer
        nama:
          example: Bank Central Asia Tbk
          type: string
        total_shares:
          example: 124275050000
          format: int64
          type: integer
        unlisted_shares:
          example: 1000000000
          format: int64
          type: integer
      type: object
    httpx.ErrorResponse:
      properties:
        error:
          type: string
      type: object
...
    httpx.Page-emiten_EmitenView:
      properties:
        data:
          items:
            $ref: '#/components/schemas/emiten.EmitenView'
          type: array
          uniqueItems: false
        limit:
          type: integer
        page:
          type: integer
        total:
          type: integer
      type: object
...
    underwriter.AllocationView:
      properties:
        nama:
          example: Mirae Asset Sekuritas
          type: string
        participant:
          description: Broker code the shares were credited to.
          example: YP
          type: string
        percentage:
          description: Share of listed_shares this tranche represents, to two decimals.
          example: 62.5
          type: number
        shares:
          description: Shares allocated.
          example: 25000
          format: int64
          type: integer
        underwriter:
          description: |-
            Broker code that took this tranche. Identical to participant, since an
            underwriter is a participant; both are present so the field a client keys
            on does not depend on knowing that.
          example: YP
          type: string
        value:
          description: 'Proceeds for this tranche: shares × ipo_price.'
          example: 25000000
          format: int64
          type: integer
      type: object
...
    underwriter.ExistingIPORequest:
      properties:
        ipo_price:
          description: |-
            Offering price per share. Must be > 0; becomes the instrument's reference
            price until it first trades.
          example: 1000
          format: int64
          type: integer
        underwriters:
          description: |-
            The syndicate. At least one entry, and the tranches must sum to exactly
            the instrument's existing listed_shares. Every member is on equal terms.
          items:
            $ref: '#/components/schemas/underwriter.UnderwriterAllocation'
          type: array
          uniqueItems: false
      required:
      - ipo_price
      - underwriters
      type: object
...
    underwriter.IPOResponse:
      properties:
        ipo_price:
          description: Offering price per share.
          example: 1000
          format: int64
          type: integer
        kode:
          example: SSIA
          type: string
        listed_shares:
          example: 40000
          format: int64
          type: integer
        nama:
          example: PT Surya Cipta Semesta Tbk
          type: string
        total_shares:
          example: 50000
          format: int64
          type: integer
        total_value:
          description: 'Total raised: listed_shares × ipo_price.'
          example: 40000000
          format: int64
          type: integer
        underwriters:
          description: The syndicate, lead first.
          items:
            $ref: '#/components/schemas/underwriter.AllocationView'
          type: array
          uniqueItems: false
        unlisted_shares:
          example: 10000
          format: int64
          type: integer
      type: object
    underwriter.UnderwriterAllocation:
      properties:
        shares:
          description: Shares allocated. Must be > 0.
          example: 25000
          format: int64
          type: integer
        underwriter:
          description: Broker code taking this tranche. Must be a registered underwriter.
          example: YP
          type: string
      required:
      - shares
      - underwriter
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      description: Static administrative key, from configuration. Guards every /api/admin route.
      in: header
      name: X-API-Key
      type: apiKey
    ParticipantKeyAuth:
      description: Per-broker key, issued by an admin and stored hashed. Guards every /api/participant route. Returned once at issue and never retrievable again.
      in: header
      name: X-Participant-Key
      type: apiKey
...
info:
  description: |-
    REST surface for the JAST core matching engine (JATS-style continuous
    matching for Bursa Efek Indonesia). The engine knows only brokers
    (participants), not individual investors, and stops at trade execution.


    There are two authentication tiers, and they never mix:

    - **Participant** (`/api/participant/*`, `/ws/participant/*`) — a broker
    authenticates with its own key, stored hashed in the database and sent as
    `X-Participant-Key`. Keys are issued and revoked per broker at runtime.
    - **Admin** (`/api/admin/*`, `/ws/admin/*`) — the single static key from
    configuration, sent as `X-API-Key`.

    Orders enter only through POST /api/participant/orders. The order book can
    be polled or streamed over WebSocket (documented, but not exercisable from
    Swagger UI, since browsers cannot set headers on a WS handshake).
  title: JAST Core API
  version: 1.0.0
openapi: 3.0.3
paths:
...
  /api/admin/emiten:
    get:
      description: |-
        Master data for every emiten, ordered by code. Price statistics are
        deliberately excluded — computing them per row would be a query per
        instrument. Use the detail endpoint for those.
      operationId: listEmiten
      parameters:
      - description: Page number, from 1
        in: query
        name: page
        schema:
          default: 1
          minimum: 1
          type: integer
      - description: Items per page, max 100
        in: query
        name: limit
        schema:
          default: 10
          maximum: 100
          minimum: 1
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.Page-emiten_EmitenView'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Missing or wrong X-API-Key
      security:
      - ApiKeyAuth: []
      summary: List listed instruments
      tags:
      - admin
    post:
      description: |-
        Registers an emiten and gives it an empty order book. It starts
        **inactive** and is *not* tradeable — `is_active` is false and every order
        against it is rejected.

        That is deliberate: a listing is an instrument existing *and* its shares
        being placed with the market. This endpoint does only the first. An
        instrument whose shares sit nowhere has nobody who could sell it and no
        offering price to quote against, so opening its book would advertise a
        market that cannot trade.

        To make it tradeable, run its offering with
        `POST /api/admin/emiten/{kode}/ipo`. That is the only way an instrument
        becomes active, and it is also where ipo_price is set — the offering price
        is decided when the offering runs, not when the instrument is registered,
        which is why this request does not take one.

        A dormant instrument is fully readable in the meantime, and is excluded
        from the composite index until it is priced.
      operationId: createEmiten
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/emiten.CreateRequest'
              description: Instrument to register
              summary: body
        description: Instrument to register
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/emiten.EmitenView'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Missing kode/nama, or listed_shares <= 0
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Missing or wrong X-API-Key
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Instrument code already exists
      security:
      - ApiKeyAuth: []
      summary: Register a new instrument (dormant)
      tags:
      - admin
  /api/admin/emiten/{kode}:
    get:
      description: |-
        The same instrument detail the participant tier reads, on the admin key —
        so oversight does not require holding a broker's credential. The admin list
        omits price statistics — computing them per row would be a query per
        instrument — so this endpoint is where they are read.

        current_price, highest_price and lowest_price are strictly trade-derived and
        are **null**, not 0, for an instrument that has never traded. reference_price
        is what the instrument is valued at and falls back to ipo_price until the
        first trade; price_source names which of the two it came from.

        value is market capitalisation (reference price × total shares), derived on
        read rather than stored, so it always reflects the latest trade.
        free_float_percentage is the publicly tradeable share of the total and
        always sums to 100 with unlisted_percentage.
      operationId: getAdminEmitenDetail
      parameters:
      - description: Emiten code
        example: BBCA
        in: path
        name: kode
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/emiten.EmitenDetail'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Missing or wrong X-API-Key
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Unknown emiten
      security:
      - ApiKeyAuth: []
      summary: Get instrument detail (admin)
      tags:
      - admin
  /api/admin/emiten/{kode}/ipo:
    post:
      description: |-
        Takes an instrument that was registered through POST /api/admin/emiten —
        and is therefore still **dormant** — public: it hands the shares to the
        syndicate and opens the order book. This is the only way an instrument
        becomes tradeable.

        The body carries no share counts. listed_shares and unlisted_shares are the
        instrument's own, fixed when it was registered — an offering decides who
        underwrites it and at what price, not how many shares the company has. The
        tranches must sum to exactly the listed_shares it already carries.

        The syndicate rules are the same as POST /api/admin/ipo:
        - each `underwriter` is a **broker code** that has been permitted to underwrite
        - the tranches must sum to **exactly** the instrument's listed_shares
        - every member is on equal terms; there is no lead

        ipo_price becomes the instrument's reference price until it first trades. An
        instrument that is already trading is rejected — a second offering would
        issue its shares twice.
      operationId: runIPOForEmiten
      parameters:
      - description: Emiten code
        example: BBNI
        in: path
        name: kode
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/underwriter.ExistingIPORequest'
              description: Offering to run
              summary: body
        description: Offering to run
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/underwriter.IPOResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Bad syndicate, shares do not sum, unknown underwriter, or emiten
            already trading
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Missing or wrong X-API-Key
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Unknown emiten
      security:
      - ApiKeyAuth: []
      summary: Run an IPO over a registered instrument
      tags:
      - admin
...
  /api/participant/emiten/{kode}:
    get:
      description: |-
        Master data plus all-time price statistics and derived valuations.

        current_price, highest_price, lowest_price and value are **null**, not 0,
        for an instrument that has never traded — no price exists yet, which is
        not the same as a price of zero.

        value is market capitalisation (current price × total shares), derived on
        read rather than stored, so it always reflects the latest trade.
        free_float_percentage is the publicly tradeable share of the total and
        always sums to 100 with unlisted_percentage.
      operationId: getEmitenDetail
      parameters:
      - description: Emiten code
        example: BBCA
        in: path
        name: kode
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/emiten.EmitenDetail'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Missing or wrong X-Participant-Key
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Unknown emiten
      security:
      - ParticipantKeyAuth: []
      summary: Get instrument detail
      tags:
      - participant
...
servers:
- description: Local development
  url: http://localhost:8080